Risk prediction Snp index construction
Will need to check the sample bim file when we construct the SNP index. SNPs that doesn't contain any genotype information should be discarded. We need to check the ordering of the p-value file and the genotype file such that to make sure the SNPs are the same in both file. e.g. Same corrdinates and same strand. Will need a way to record the strand flipping